refactor. name을 대체하는 Name 클래스 추가
