+++ src/main/java/racingcargame/controller/CarRace.java
@@ -1,0 +1,5 @@
+package racingcargame.controller;
+
+public class CarRace {
+
+}
